Movign to the LAMMPSlib from ASE
