+++ default_constants.py
@@ -73,6 +73,31 @@ default_charges = {
     "H": 1.2
 }
 
+bks_params = {
+    # buck/coul/long: (A_ij, eV; b_ij, A^-1, c_ij, eV*A^6)
+    ("Si", "Si"): (834.40, 1/0.29, 0.0),
+    ("O",  "O" ): (1388.7730, 2.76000,175.0000),
+    ("Si", "O" ): (45296.72, 1/0.161,46.1395),
+    ("Si", "Al"): (646.67, 1/0.120, 0.0),
+    ("O",  "Al"): (28287.00, 1/0.172, 34.7600),
+    ("Al", "Al"): (351.94, 1/0.360, 0.0),
+}
+# lj/cut: (epsilon, sigma, sigma)
+bks_lj = {
+    ("Si", "Si"): ( 0.0, 0.0, 0.0),
+    ("O",  "O" ): (2.6, 1.6, 1.6),
+    ("Si", "O" ): (2.0, 1.2, 1.2),
+    ("Si", "Al"): (0.0, 0.0, 0.0),
+    ("O",  "Al"): (2.3, 1.4, 1.4),
+    ("Al", "Al"): ( 0.0, 0.0, 0.0),
+}
+# charges
+bks_charges = {
+    "Si":  2.4,
+    "O" : -1.2,
+    "Al":  1.8,
+}
+
 NORM_COORDINATION = {"Si": 4,"Al":4, "O": 2, "H": 1}
 
 #bond lenghst for satuarion routine
