+++ requirements.txt
@@ -4,3 +4,4 @@ matplotlib~=3.10.0
 tqdm~=4.67.1
 torch~=2.7.0
 mace-torch~= 0.3.13
+lammps~=2024.8.29.3.0
